+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Resolved issues:
 
-- CE-1183 Fixed filter in the recordings page. The user is able to select the month of January
+- CE-1230 Fix the issue with filtered recordings by species
 
 ## v3.0.34 - August XX, 2021
 
@@ -17,6 +17,7 @@ New features:
 Resolved issues:
 
 - CE-1197 Change wording when choosing filename format in the uploads page
+- CE-1183 Fixed filter in the recordings page. The user is able to select the month of January
 
 ## v3.0.33-hotfix.0 - August 03, 2021